requirements.txt
opencv_python==4.5.3.56
numpy==1.21.2
python_telegram_bot==13.7